+++ tests/tools/test_json_schema_validator.py
@@ -1,6 +1,7 @@
 from collections import deque
 
 from arche.tools.json_schema_validator import JsonSchemaValidator as JSV
+import pandas as pd
 import pytest
 
 
@@ -64,25 +65,22 @@ def test_format_validation_message(
     )
 
 
-@pytest.mark.parametrize(
-    "schema_source, expected_schema",
-    [
-        (
-            {"$schema": "http://json-schema.org/draft-07/schema"},
-            {"$schema": "http://json-schema.org/draft-07/schema"},
-        )
-    ],
-)
-def test_validator(schema_source, expected_schema):
-    jsv = JSV(schema_source)
-    assert jsv.schema == expected_schema
+def test_validate():
+    jsv = JSV({"properties": {"NAME": {"type": "string"}}})
+    jsv.validate(pd.DataFrame([{"NAME": None, "udkey": "0"}]))
+    assert jsv.errors == {"NAME is not of type 'string'": {"0"}}
 
 
-@pytest.mark.parametrize(
-    "items, expected_errors",
-    [([{"_key": "0"}, {"_key": "1"}], {"data must be number": {"0", "1"}})],
-)
-def test_fast_validate(items, expected_errors):
+def test_fast_validate():
     jsv = JSV({"type": "number"})
-    jsv.fast_validate(items)
-    assert jsv.errors == expected_errors
+    jsv.fast_validate(pd.DataFrame([{"udkey": "0"}, {"udkey": "1"}]))
+    assert jsv.errors == {"data must be number": {"0", "1"}}
+
+
+def test_run():
+    schema = {"properties": {"NAME": {"type": "string"}}}
+    jsv = JSV(schema)
+    assert jsv.schema == schema
+    jsv.run(pd.DataFrame([{"NAME": None, "_key": "0", "__DEBUG": None}]), fast=False)
+
+    assert jsv.errors == {"NAME is not of type 'string'": {"0"}}